[Refactor] 클래스 구조 분리, 불필요한 파라미터 검증 로직 제거, 캐시 어노테이션 위치 변경, 테스트 수정

+++ src/main/java/com/musinsa/watcher/web/dto/TodayDiscountedProductDto.java
@@ -1,6 +1,6 @@
 package com.musinsa.watcher.web.dto;
 
-import com.musinsa.watcher.domain.price.TodayDiscountProduct;
+import com.musinsa.watcher.domain.product.discount.TodayDiscountProduct;
 import java.io.Serializable;
 import java.time.LocalDate;
 import lombok.Getter;
@@ -15,7 +15,7 @@ public class TodayDiscountedProductDto implements Serializable {
   private final LocalDate modifiedDate;
   private final int price;
   private final long discount;
-  private final float percent;
+  private final double percent;
 
   public TodayDiscountedProductDto(TodayDiscountProduct entity) {
     this.productId = entity.getProduct().getProductId();
